printf around the corner

+++ LVL1/ft_printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: hdrabi <hdrabi@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/06 16:09:19 by hdrabi            #+#    #+#             */
-/*   Updated: 2021/11/22 11:19:38 by hdrabi           ###   ########.fr       */
+/*   Updated: 2021/11/30 17:39:13 by hdrabi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ void	ft_striteri(char *s, void (*f)(unsigned int, char *));
 void	ft_putchar_fd(char c, int fd);
 void	ft_putstr_fd(char *s, int fd);
 void	ft_putendl_fd(char *s, int fd);
-void	ft_putnbr_fd(int n, int fd);
+void	ft_putnbr_fd(long long n, int fd);
 typedef struct s_list
 {
 	void			*content;
